+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.10...HEAD) - TBD
 
+### :rocket: Added
+
+- Class-based checks that run per response and once after the whole run. [#1147](https://github.com/schemathesis/schemathesis/issues/1147), [#1689](https://github.com/schemathesis/schemathesis/issues/1689)
+
 ## [4.21.10](https://github.com/schemathesis/schemathesis/compare/v4.21.9...v4.21.10) - 2026-06-20
 
 ### :bug: Fixed